Adding wrapper generation and build instructions and etc.
Added instructions for generating wrapper code and creating
dll for them.

Adapted vs project to the changes and created an example. Please note
that I've created DLLs for x64 architecture so it can only be used in
that configuration. If you want x86 then you need to compile kuzu for
that archtecture.